Remote client doesn't store its type
With OOP, it is bad design to store the type of an object in
a member variable. On top of that, it was more than confusing
that vvRemoteClient was responsible to send the type over the
socket, while vserver was responsible to receive it. Changed
this so that the tool is responsible to communicate the type.
Anyway, this is probably still subject to change in the future