Fix cuda::Symbol.
The current implementation is incompatible with #include <cuda_runtime.h>.
This header defines a function template which takes the symbol by reference and forwards the address of this reference to the C API cudaMemcpyToSymbol. Since Name() returns a temporary, this results in an invalid symbol address.
So make sure the C API function is called.

Use VV_CUDA_INIT_SYMBOL to initialize symbols.

(NOTE: Not tested with CUDA 4.x)