Fix tricubic filtering with b-splines
We validated this version using 1-d and 2-d plots and are quite
confident that the filter now works as intended.

Note that this is a filter, which does not *interpolate* the volume
and thus in general will not reconstruct it at the voxels. For that,
one would have to to initially prefilter the volume (WIP).

Changes by ypercan and zellmans.